Refactor: Move common classes to utils package (feat/webhook)
Moved classes from the `park.bumsiku.common` package to the
`park.bumsiku.utils` package. This change improves the
organization of the project by grouping utility classes
together.  Affected classes include:

- ArgumentValidator
- ArgumentValidatorImpl
- ApplicationStartupLogger
- DiscordWebhookCreator
- GlobalExceptionHandler
- LoggingFilter

The logback-spring.xml and related test classes were also
updated to reflect this change.  Import statements in
controllers and tests were modified accordingly.
