+++ src/main/java/park/bumsiku/utils/ApplicationStartupLogger.java
@@ -1,4 +1,4 @@
-package park.bumsiku.common;
+package park.bumsiku.utils;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
