+++ src/main/resources/logback-spring.xml
@@ -28,7 +28,7 @@
         <appender-ref ref="FILE"/>
     </logger>
 
-    <logger name="park.bumsiku.common.LoggingFilter" level="INFO" additivity="false">
+    <logger name="park.bumsiku.utils.LoggingFilter" level="INFO" additivity="false">
         <appender-ref ref="CONSOLE"/>
         <appender-ref ref="FILE"/>
     </logger>
