+++ src/test/java/park/bumsiku/controller/AdminControllerTest.java
@@ -10,7 +10,7 @@ import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
-import park.bumsiku.common.ArgumentValidator;
+import park.bumsiku.utils.ArgumentValidator;
 import park.bumsiku.config.Security;
 import park.bumsiku.domain.dto.request.CreateCategoryRequest;
 import park.bumsiku.domain.dto.request.CreatePostRequest;
