+++ src/test/java/park/bumsiku/controller/PublicControllerTest.java
@@ -8,7 +8,7 @@ import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
-import park.bumsiku.common.ArgumentValidator;
+import park.bumsiku.utils.ArgumentValidator;
 import park.bumsiku.config.Security;
 import park.bumsiku.domain.dto.request.CommentRequest;
 import park.bumsiku.domain.dto.response.*;
