+++ src/test/java/park/bumsiku/utils/ArgumentValidatorImplTest.java
@@ -1,4 +1,4 @@
-package park.bumsiku.common;
+package park.bumsiku.utils;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
